+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `selectors.sizeRatioPercent` metric
 - Added `declarations.sizeRatioPercent` metric
 - Added `dataUris.sizeRatioPercent` metric
+- Added `selectors.attributesUsage` metric
+- Added `attributesUsage` flag
 
 ### Changed